api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "line-bot-sdk>=3.21.0",
    "openai>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
